data_ingestion/ingestion_requirements.txt
PyYAML>=6.0
google-cloud-pubsub==2.31.1
websockets==15.0.1
requests==2.32.5
python-dotenv==1.1.1
aiohttp==3.12.15
